# src/docs/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: HealthNest API
  version: 1.0.0
  description: HealthNest Backend REST API.
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development
  - url: https://healthnest-juho.onrender.com/api/v1
    description: Production (Render)
tags:
  - name: Authentication
  - name: Users
  - name: Health
  - name: Medications
  - name: Appointments
paths:
  /health:
    get:
      tags:
        - Health
      summary: API Health Check
      description: |
        Checks whether the HealthNest API is running.

        This endpoint is commonly used by:

          • Load balancers
          • Monitoring systems
          • Kubernetes probes
          • Mobile/Web applications
      operationId: getHealth
      responses:
        '200':
          description: API is running successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                success: true
                message: API is running.
                data:
                  status: healthy
                errors: null
                meta: null
                timestamp: '2026-07-20T10:15:30.456Z'
  /users/register:
    post:
      tags:
        - Users
      summary: Register New User
      description: |
        Creates a new HealthNest user account.

        The email address must be unique.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              firstName: Promise
              lastName: Okere
              email: promise@example.com
              password: Password@123
      responses:
        '201':
          description: User registered successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterSuccessResponse'
              example:
                success: true
                message: User registered successfully.
                data:
                  id: 686eb58c1942547e95ce5e61
                  firstName: Promise
                  lastName: Okere
                  email: promise@example.com
                  role: user
                  createdAt: '2026-07-20T14:32:10.000Z'
                  updatedAt: '2026-07-20T14:32:10.000Z'
                errors: null
                meta: null
                timestamp: '2026-07-20T14:32:10.000Z'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: |
        Authenticates a user and returns
        an Access Token and Refresh Token.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: promise@example.com
              password: Password@123
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                success: true
                message: Login successful.
                data:
                  user:
                    id: 686eb58c1942547e95ce5e61
                    firstName: Promise
                    lastName: Okere
                    email: promise@example.com
                    role: user
                  accessToken: eyJhbGc...
                  refreshToken: eyJhbGc...
                errors: null
                meta: null
                timestamp: '2026-07-20T14:32:10.000Z'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get Current User
      description: |
        Returns the currently authenticated user.
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: Authenticated successfully.
                data:
                  id: 686eb58c1942547e95ce5e61
                  firstName: Promise
                  lastName: Okere
                  email: promise@example.com
                  role: user
                errors: null
                meta: null
                timestamp: '2026-07-20T14:32:10.000Z'
  /medications:
    post:
      tags:
        - Medications
      summary: Create Medication
      description: |
        Creates a medication schedule for the authenticated user.
      operationId: createMedication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMedicationRequest'
            example:
              name: Paracetamol
              dosage: 500mg
              frequency: Twice Daily
              reminderTime:
                - '08:00'
                - '20:00'
              startDate: '2026-07-26'
              endDate: '2026-08-05'
              instructions: Take after meals
      responses:
        '201':
          description: Medication created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medication'
              example:
                success: true
                message: Medication created successfully.
                data:
                  id: 686eb58c1942547e95ce5e61
                  name: Paracetamol
                  dosage: 500mg
                  frequency: Twice Daily
                  reminderTime:
                    - '08:00'
                    - '20:00'
                  status: Active
                  startDate: '2026-07-26'
                  endDate: '2026-08-05'
                  instructions: Take after meals
                errors: null
                meta: null
                timestamp: '2026-07-26T18:30:00.000Z'
    get:
      tags:
        - Medications
      summary: Get User Medications
      description: Returns all medications belonging to the authenticated user.
      operationId: getMedications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Medication list retrieved successfully.
  /appointments:
    post:
      tags:
        - Appointments
      summary: Create Appointment
      description: |
        Creates a medical appointment for the authenticated user.
      operationId: createAppointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
            example:
              hospitalName: Lagoon Hospital
              doctorSpeciality: Cardiologist
              doctorName: Dr. Michael Ade
              appointmentDate: '2026-08-10'
              appointmentTime: '09:30'
              note: Annual heart checkup
      responses:
        '201':
          description: Appointment created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              example:
                success: true
                message: Appointment created successfully.
                data:
                  id: 686eb58c1942547e95ce5e61
                  hospitalName: Lagoon Hospital
                  doctorSpeciality: Cardiologist
                  doctorName: Dr. Michael Ade
                  appointmentDate: '2026-08-10'
                  appointmentTime: '09:30'
                  note: Annual heart checkup
                errors: null
                meta: null
                timestamp: '2026-07-26T18:30:00.000Z'
    get:
      tags:
        - Appointments
      summary: Get User Appointments
      description: Returns all appointments belonging to the authenticated user.
      operationId: getAppointments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Appointment list retrieved successfully.
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Request completed successfully.
        data:
          nullable: true
        errors:
          nullable: true
        meta:
          nullable: true
        timestamp:
          type: string
          format: date-time
          example: '2026-07-20T13:15:25.345Z'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed.
        data:
          nullable: true
          example: null
        errors:
          type: object
          example:
            email:
              - Email is required.
            password:
              - Password must be at least 8 characters.
        meta:
          nullable: true
          example: null
        timestamp:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
      properties:
        firstName:
          type: string
          example: Promise
        lastName:
          type: string
          example: Okere
        email:
          type: string
          format: email
          example: promise@example.com
        password:
          type: string
          format: password
          example: Password@123
    User:
      type: object
      properties:
        id:
          type: string
          example: 686eb58c1942547e95ce5e61
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          example: user
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: promise@example.com
        password:
          type: string
          format: password
          example: Password@123
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Login successful.
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
            refreshToken:
              type: string
        errors:
          nullable: true
          example: null
        meta:
          nullable: true
          example: null
        timestamp:
          type: string
          format: date-time
    Medication:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Medication created successfully.
        data:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            dosage:
              type: string
            frequency:
              type: string
            reminderTime:
              type: array
              items:
                type: string
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
            instructions:
              type: string
            status:
              type: string
              example: Active
        errors:
          nullable: true
        meta:
          nullable: true
        timestamp:
          type: string
          format: date-time
    CreateMedicationRequest:
      type: object
      required:
        - name
        - dosage
        - startDate
      properties:
        name:
          type: string
          example: Paracetamol
        dosage:
          type: string
          example: 500mg
        frequency:
          type: string
          example: Twice Daily
        reminderTime:
          type: array
          items:
            type: string
          example:
            - '08:00'
            - '20:00'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        instructions:
          type: string
    UpdateMedicationRequest:
      allOf:
        - $ref: '#/components/schemas/CreateMedicationRequest'
    Appointment:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Appointment created successfully.
        data:
          type: object
          properties:
            id:
              type: string
            hospitalName:
              type: string
            doctorSpeciality:
              type: string
            doctorName:
              type: string
            appointmentDate:
              type: string
              format: date
            appointmentTime:
              type: string
            note:
              type: string
        errors:
          nullable: true
        meta:
          nullable: true
        timestamp:
          type: string
          format: date-time
    CreateAppointmentRequest:
      type: object
      required:
        - hospitalName
        - appointmentDate
        - appointmentTime
      properties:
        hospitalName:
          type: string
          example: Lagoon Hospital
        doctorSpeciality:
          type: string
          example: Cardiologist
        doctorName:
          type: string
          example: Dr. Michael Ade
        appointmentDate:
          type: string
          format: date
        appointmentTime:
          type: string
          example: '09:30'
        note:
          type: string
    UpdateAppointmentRequest:
      allOf:
        - $ref: '#/components/schemas/CreateAppointmentRequest'
    HealthData:
      type: object
      properties:
        status:
          type: string
          example: healthy
    HealthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: API is running.
        data:
          $ref: '#/components/schemas/HealthData'
        errors:
          nullable: true
          example: null
        meta:
          nullable: true
          example: null
        timestamp:
          type: string
          format: date-time
    RegisterSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: User registered successfully.
        data:
          $ref: '#/components/schemas/User'
        errors:
          nullable: true
          example: null
        meta:
          nullable: true
          example: null
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT